libnetwork: Move bridge.IsV6Listenable to netutils
